refactor(config): drop JSON admin-file paths and widen Port to int32
AdminFileStore's LoadJsonFile / SaveJsonFile helpers and the five
per-file JSON migration branches in LoadAll are deleted outright.
The store only reads the per-key TOML format now; nlohmann/json and
<filesystem> drop out of the includes. Callers that relied on the
legacy JSON round-trip need to regenerate their admin files by hand.

ServerConfig::Port widens from uint16_t to int32_t so command-line
parsing and the config round-trip don't truncate out-of-range
values before validation. The underlying TOML node is already i64.

DedicatedProperties loses a stale ImportInto comment that pointed
at a long-deleted migration helper.

+++ Axiom/src/Axiom/Config/ServerConfig.cpp
@@ -45,8 +45,8 @@ std::string ServerConfig::BindAddress() const {
 	return m_Root["network"]["bind-address"].value_or<std::string>("0.0.0.0");
 }
 
-uint16_t ServerConfig::Port() const {
-	return static_cast<uint16_t>(m_Root["network"]["port"].value_or(25565));
+int32_t ServerConfig::Port() const {
+	return static_cast<int32_t>(m_Root["network"]["port"].value_or(25565));
 }
 
 std::string ServerConfig::ServerName() const {
